rename module for spring redis data 2.x
