--- opentracing-redis-spring/src/test/java/io/opentracing/contrib/redis/spring/processor/RedisConnectionFactoryBeanPostProcessor.java
@@ -1,39 +1,0 @@
-/*
- * Copyright 2017-2019 The OpenTracing Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
- */
-package io.opentracing.contrib.redis.spring.processor;
-
-import io.opentracing.Tracer;
-import io.opentracing.contrib.redis.spring.connection.TracingRedisConnectionFactory;
-import org.springframework.beans.BeansException;
-import org.springframework.beans.factory.config.BeanPostProcessor;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
-import org.springframework.stereotype.Component;
-
-@Component
-public class RedisConnectionFactoryBeanPostProcessor implements BeanPostProcessor {
-
-    private Tracer tracer;
-
-    public RedisConnectionFactoryBeanPostProcessor(Tracer tracer) {
-        this.tracer = tracer;
-    }
-
-    @Override
-    public Object postProcessAfterInitialization(Object bean, String beanName) throws BeansException {
-        if (bean instanceof RedisConnectionFactory) {
-            bean = new TracingRedisConnectionFactory((RedisConnectionFactory) bean, false, tracer);
-        }
-        return bean;
-    }
-}
