--- opentracing-redis-spring/src/test/java/io/opentracing/contrib/redis/spring/processor/RedisConnectionFactoryBeanPostProcessorConfiguration.java
@@ -1,29 +1,0 @@
-/*
- * Copyright 2017-2019 The OpenTracing Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
- */
-package io.opentracing.contrib.redis.spring.processor;
-
-import io.opentracing.Tracer;
-import io.opentracing.mock.MockTracer;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@ComponentScan("io.opentracing.contrib.redis.spring.processor")
-public class RedisConnectionFactoryBeanPostProcessorConfiguration {
-    @Bean
-    public Tracer tracer() {
-        return new MockTracer();
-    }
-}
